Tabu project update

+++ Tabu/Services/Implements/LanguageService.cs
@@ -4,37 +4,46 @@ using Microsoft.EntityFrameworkCore;
 using Tabu.DAL;
 using Tabu.DTOs.Languages;
 using Tabu.Entities;
+using Tabu.Exceptions.Language;
 using Tabu.Services.Abstracts;
 
 namespace Tabu.Services.Implements;
 
 public class LanguageService(TabuDbContext _context, IMapper _mapper) : ILanguageService
 {
-    public IEnumerable<LanguageGetDto> GetAsync()
+    public async Task<IEnumerable<LanguageGetDto>> GetAllAsync()
     {
-        var data = _context.Languages.ToList();
-        var lang = data.Select(x => _mapper.Map<LanguageGetDto>(x)).ToList();
-        return lang;
+        var lang = await _context.Languages.ToListAsync();
+        return _mapper.Map<IEnumerable<LanguageGetDto>>(lang);
     }
     public async Task CreateAsync(LanguageCreateDto dto)
     {
+        if (await _context.Languages.AnyAsync(x => x.Code == dto.Code))
+            throw new LanguageExistException();
         var lang = _mapper.Map<Language>(dto);
         await _context.AddAsync(lang);
         await _context.SaveChangesAsync();
     }
     public async Task UpdateAsync(string code,LanguageUpdateDto dto)
     {
-        var lang = _mapper.Map<Language>(dto);
-        var data = await _context.Languages.Where(x => x.Code == code).FirstOrDefaultAsync();
-        data.Code = dto.Code;
-        data.Name = dto.Name;
-        data.Icon = dto.IconUrl;
+        var data = await _getByCode(code);
+        if (data == null) throw new LanguageNotFoundException();
+        _mapper.Map(dto, data);
         await _context.SaveChangesAsync();
     }
     public async Task DeleteAsync(string code)
     {
-        var data = await _context.Languages.Where(x => x.Code == code).FirstOrDefaultAsync();
+        if (await _context.Languages.AnyAsync(x => x.Code == code))
+            throw new LanguageExistException();
+        var data = await _getByCode(code);
         _context.Languages.Remove(data);
         await _context.SaveChangesAsync();
     }
+
+    public Task<LanguageGetDto> GetByCode(string code)
+    {
+        throw new NotImplementedException();
+    }
+    async Task<Language?> _getByCode(string code)
+    => await _context.Languages.FindAsync(code);
 }
